move code with hid from keymap to rev2
