--- keyboards/ergohaven/planeta/keymaps/v2/lv_conf.h
@@ -1,12 +1,0 @@
-#pragma once
-
-#include_next <lv_conf.h>
-
-#undef LV_FONT_DEFAULT
-#define LV_FONT_DEFAULT &lv_font_montserrat_20
-
-#undef LV_FONT_MONTSERRAT_20
-#define LV_FONT_MONTSERRAT_20 1
-
-#undef LV_FONT_MONTSERRAT_28
-#define LV_FONT_MONTSERRAT_28 0
