+++ keyboards/ergohaven/planeta/keymaps/v2/rules.mk
@@ -1,4 +1,1 @@
 VIAL_ENABLE = yes
-RAW_ENABLE = yes
-
-SRC += ./hid_display.c
