+++ keyboards/ergohaven/planeta/rev2/lv_conf.h
@@ -14,6 +14,9 @@
 #undef LV_FONT_MONTSERRAT_14
 #define LV_FONT_MONTSERRAT_14 0
 
+#undef LV_FONT_MONTSERRAT_28
+#define LV_FONT_MONTSERRAT_28 0
+
 #undef LV_FONT_MONTSERRAT_20
 #define LV_FONT_MONTSERRAT_20 1
