+++ keyboards/ergohaven/planeta/rev2/rev2.c
@@ -1,17 +1,10 @@
-#include "rev2.h"
 #include "display.h"
 
-static bool display_enabled;
-
-/* public function to be used in keymaps */
-bool is_display_enabled(void) {
-    return display_enabled;
-}
 
 /* Caps Lock processing */
 bool led_update_kb(led_t led_state) {
     bool res = led_update_user(led_state);
-    if (res && display_enabled) {
+    if (res && is_display_enabled()) {
         display_process_caps(led_state.caps_lock);
     }
 
@@ -19,7 +12,7 @@ bool led_update_kb(led_t led_state) {
 }
 
 void housekeeping_task_kb(void) {
-    if (display_enabled) {
+    if (is_display_enabled()) {
         display_housekeeping_task();
     }
 
@@ -27,7 +20,6 @@ void housekeeping_task_kb(void) {
 }
 
 void keyboard_post_init_user(void) {
-    display_enabled = false;
-    display_enabled = display_init_kb();
+    display_init_kb();
     writePinHigh(GP18);
 }
