--- keyboards/ergohaven/planeta/rev2/rev2.h
@@ -1,3 +1,0 @@
-#pragma once
-
-bool is_display_enabled(void);
