+++ keyboards/ergohaven/planeta/rev2/rules.mk
@@ -6,5 +6,8 @@ QUANTUM_PAINTER_DRIVERS += st7789_spi
 QUANTUM_PAINTER_LVGL_INTEGRATION = yes
 RGBLIGHT_ENABLE = yes
 
+RAW_ENABLE = yes
+
 SRC += lvgl_helpers.c \
-	   display.c
+	   display.c \
+	   hid_display.c
